Removing Errors
Updated the test_mdamath.py and test_util.py
